Add model evaluation plots and update data processing pipeline
